Get hourly weather by facility id

+++ src/main/java/cz/expertkom/sportFacilities/dto/FacilityDto.java
@@ -18,8 +18,8 @@ public class FacilityDto {
     private int managerId; // FK
     private Boolean indoor;
     private Boolean lighting;
-    private double longitude;
-    private double latitude;
+    private float longitude;
+    private float latitude;
     private String photos;
     private BigDecimal rating;
     private String description;
