+++ src/main/java/cz/expertkom/sportFacilities/model/Facility.java
@@ -37,8 +37,8 @@ public class Facility {
     private Integer lastEditedBy;
     private Boolean indoor;
     private Boolean lighting;
-    private double longitude;
-    private double latitude;
+    private float longitude;
+    private float latitude;
     private String photos;
     private BigDecimal rating;
     private String description;
